refactor: query package
