+++ packages/query/index.ts
@@ -1,10 +1,9 @@
 export * from './src/bitcoin/address/address.utils';
-export * from './src/bitcoin/address/transactions-by-address.hooks';
 export * from './src/bitcoin/address/transactions-by-address.query';
-export * from './src/bitcoin/address/utxos-by-address.hooks';
 export * from './src/bitcoin/address/utxos-by-address.query';
 export * from './src/bitcoin/balance/btc-blockbook-utxos';
-export * from './src/bitcoin/balance/btc-taproot-balance.hooks';
+export * from './src/bitcoin/balance/btc-taproot-balance.query';
+export * from './src/bitcoin/balance/balance.utils';
 export * from './src/bitcoin/clients/best-in-slot';
 export * from './src/bitcoin/clients/bitcoin-client';
 export * from './src/bitcoin/fees/fee-estimates.query';
@@ -13,7 +12,7 @@ export * from './src/bitcoin/ordinals/inscription-text-content.query';
 export * from './src/bitcoin/ordinals/inscription.query';
 export * from './src/bitcoin/ordinals/inscription.utils';
 export * from './src/bitcoin/ordinals/inscriptions-by-param.query';
-export * from './src/bitcoin/ordinals/inscriptions.hooks';
+export * from './src/bitcoin/ordinals/ordinals.utils';
 export * from './src/bitcoin/ordinals/inscriptions.query';
 export * from './src/bitcoin/ordinalsbot-client';
 export * from './src/bitcoin/runes/runes-outputs-by-address.query';
@@ -27,8 +26,8 @@ export * from './src/common/compliance-checker/compliance-checker.query';
 export * from './src/common/compliance-checker/compliance-checker.constants';
 export * from './src/common/compliance-checker/compliance-checker.types';
 export * from './src/common/alex-sdk/alex-sdk-swappable-currency.query';
-export * from './src/common/alex-sdk/alex-sdk.hooks';
-export * from './src/common/market-data/market-data.hooks';
+export * from './src/common/alex-sdk/alex-sdk.utils';
+export * from './src/common/market-data/market-data.utils';
 export * from './src/common/market-data/market-data.query';
 export * from './src/common/market-data/vendors/binance-market-data.query';
 export * from './src/common/market-data/vendors/coincap-market-data.query';
@@ -38,7 +37,7 @@ export * from './src/query-config';
 export * from './src/query-prefixes';
 export * from './src/rate-limiter/best-in-slot-limiter';
 export * from './src/rate-limiter/hiro-rate-limiter';
-export * from './src/stacks/balance/account-balance.hooks';
+export * from './src/stacks/balance/account-balance.utils';
 export * from './src/stacks/balance/account-balance.query';
 export * from './src/stacks/bns/bns.query';
 export * from './src/stacks/bns/bns.utils';
@@ -55,7 +54,7 @@ export * from './src/stacks/nonce/account-nonces.utils';
 export * from './src/stacks/sip10/sip10-tokens.utils';
 export * from './src/stacks/stacks-client';
 export * from './src/stacks/stx20-api-types';
-export * from './src/stacks/stx20/stx20-tokens.hooks';
+export * from './src/stacks/stx20/stx20-tokens.utils';
 export * from './src/stacks/stx20/stx20-tokens.query';
 export * from './src/stacks/token-metadata/fungible-tokens/fungible-token-metadata.query';
 export * from './src/stacks/token-metadata/non-fungible-tokens/non-fungible-token-holdings.query';
