--- packages/query/src/bitcoin/address/transactions-by-address.hooks.ts
@@ -1,18 +1,0 @@
-import type { BitcoinTx } from '@leather.io/models';
-import { sumNumbers } from '@leather.io/utils';
-
-export function calculateOutboundPendingTxsValue(pendingTxs: BitcoinTx[], address: string) {
-  // sum all inputs
-  const sumInputs = sumNumbers(pendingTxs.flatMap(tx => tx.vin.map(input => input.prevout.value)));
-
-  // get all outputs that are sent back to the address
-  const returnedOutputChangeValues = pendingTxs
-    .flatMap(tx => tx.vout.map(output => output))
-    .filter(v => v.scriptpubkey_address === address)
-    .flatMap(output => output.value);
-
-  // sum all filtered outputs
-  const sumOutputs = sumNumbers(returnedOutputChangeValues);
-
-  return sumInputs.minus(sumOutputs).toNumber();
-}
