+++ packages/query/src/bitcoin/address/utxos-by-address.spec.tsx
@@ -1,6 +1,6 @@
 import { mockInscriptionsList } from '../../../test/mock-inscriptions';
 import { mockUtxos } from '../../../test/mock-utxos';
-import { filterUtxosWithInscriptions } from './utxos-by-address.hooks';
+import { filterUtxosWithInscriptions } from './address.utils';
 
 describe(filterUtxosWithInscriptions, () => {
   test('that it filters out utxos with inscriptions so they are not spent', () => {
