+++ packages/query/src/bitcoin/balance/btc-taproot-balance.query.ts
@@ -1,30 +1,10 @@
 import type { UseQueryResult } from '@tanstack/react-query';
 
-import { Inscription } from '@leather.io/models';
-import { createMoney, isDefined, sumNumbers } from '@leather.io/utils';
+import { isDefined } from '@leather.io/utils';
 
-import { UtxoWithDerivationPath } from '../../../types/utxo';
-import { filterUtxosWithInscriptions } from '../address/utxos-by-address.hooks';
 import { BestInSlotInscriptionByXpubResponse } from '../clients/best-in-slot';
 import { createBestInSlotInscription } from '../ordinals/inscription.utils';
-import { findInscriptionsOnUtxo } from '../ordinals/inscriptions.hooks';
-
-const RETRIEVE_UTXO_DUST_AMOUNT = 10000;
-
-export function filterUninscribedUtxosToRecoverFromTaproot(
-  utxos: UtxoWithDerivationPath[],
-  inscriptions: Inscription[]
-) {
-  const filteredUtxosList = utxos
-    .filter(utxo => utxo.status.confirmed)
-    .filter(utxo => utxo.value > RETRIEVE_UTXO_DUST_AMOUNT);
-
-  return filteredUtxosList.filter(filterUtxosWithInscriptions(inscriptions));
-}
-
-export function utxosToBalance(utxos: UtxoWithDerivationPath[]) {
-  return createMoney(sumNumbers(utxos.map(utxo => Number(utxo.value))), 'BTC');
-}
+import { findInscriptionsOnUtxo } from '../ordinals/ordinals.utils';
 
 export function combineInscriptionResults(
   queries: UseQueryResult<BestInSlotInscriptionByXpubResponse, Error>[]
