+++ packages/query/src/common/alex-sdk/alex-sdk-swappable-currency.query.ts
@@ -1,4 +1,4 @@
-import { alex } from './alex-sdk.hooks';
+import { alex } from './alex-sdk.utils';
 
 const queryOptions = {
   refetchOnMount: true,
