+++ packages/query/src/common/market-data/market-data.utils.ts
